requirements.txt
pandas 
numpy
scikit-learn
flask
joblib
matplotlib
seaborn
azure-storage-blob
azure-identity
pyyaml